feat: make FAQ component reusable
